iss ue was .pls files would open but .m3u files wouldn't. This is now fixed but codes needs refactoring to tidy it up.

+++ pyradio.py
@@ -9,10 +9,12 @@ import datetime as d
 # below is a list of URL/m3u/plsd files to play
 urls = [
     'C:/PyRadio/jassfm.pls',
-    # 'file:///C:/PyRadio/heartxmas.m3u',
-    # "C:\PyRadio\capitalxtra.m3u",
-    # 'C:/PyRadio/virgin.m3u',
-    'C:/PyRadio/kerrang.pls'
+    #'C:/PyRadio/heartxmas.m3u',
+    "C:/PyRadio/capitalxtra.m3u",
+    'C:/PyRadio/virgin.m3u',
+    'C:/PyRadio/kerrang.pls',
+    'C:/PyRadio/bbc_radio_one.m3u8',
+    'C:/PyRadio/captialfm.m3u'
 
 
     # 'file:///C:/PyRadio/magic.pls',
@@ -39,24 +41,49 @@ def radio_player():
             playlist_url = urls[current_index]
             with open(playlist_url, "r") as radio_file:
                 radio_content = radio_file.read()
+            if "m3u" in playlist_url:
 
-            stream_url = None
-            for line in radio_content.splitlines():
-                if line.startswith("File1="):
-                    stream_url = line[len("File1="):]
-                    print(stream_url)
-                    break
+                radio_lines = radio_content.splitlines()
+                radio_entries = [line for line in radio_lines if line and not line.startswith("#")]
 
-            if stream_url:
-                vlc_process = subprocess.Popen([vlc_path, "--playlist-autostart", stream_url])
-                time.sleep(5)  # Sleep for 30 minutes (1800 seconds)
-                vlc_process.terminate()  # Terminate the current VLC instance
-                if (current_index == (len(urls)-1)):
-                    current_index = 0
-                else:
-                    current_index = (current_index + 1)
-            else:
-                print("Stream URL not found in playlist.")
+                #stream_url = None
+                #for line in radio_content.splitlines():
+                #    if line.startswith("File1="):
+                 #       stream_url = line[len("File1="):]
+                  #      print(stream_url)
+                   #     break
+
+                if radio_entries:
+
+                    stream_url = radio_entries[0]
+                    vlc_process = subprocess.Popen([vlc_path, "--playlist-autostart", stream_url])
+                    time.sleep(15)  # Sleep for 30 minutes (1800 seconds)
+                    vlc_process.terminate()  # Terminate the current VLC instance
+                    if (current_index == (len(urls)-1)):
+                        current_index = 0
+                    else:
+                        current_index = (current_index + 1)
+
+            if "pls" in playlist_url:
+                stream_url = None
+                for line in radio_content.splitlines():
+                    if line.startswith("File1="):
+                        stream_url = line[len("File1="):]
+                        print(stream_url)
+                        break
+
+                if stream_url:
+                    vlc_process = subprocess.Popen([vlc_path, "--playlist-autostart", stream_url])
+                    time.sleep(15)  # Sleep for 30 minutes (1800 seconds)
+                    vlc_process.terminate()  # Terminate the current VLC instance
+                    if (current_index == (len(urls) - 1)):
+                        current_index = 0
+                    else:
+                        current_index = (current_index + 1)
+            #else:
+
+
+             #   print("Stream URL not found in playlist.")
 
 
 
@@ -77,14 +104,6 @@ def radio_player():
                   #  time.sleep(19)
 
 radio_player()
-#schedule.every().at("09:00").do(radio_player())
-
-
-#while True:
-    # Checks whether a scheduled task
-    # is pending to run or not
- #   schedule.run_pending()
-  #  time.sleep(1)
